+++ world/rough.c
@@ -18,7 +18,7 @@ int main(void)
 
     printf("%i\n", no_of_digits(card_number));
     printf("%i\n", peter_lunh_sum(card_number));
-    printf("%li\n", first_two_digits(card_number));
+    printf("%i\n", first_two_digits(card_number));
 }
 
 // peter_lunh_sum means sum according to Lunh's algorithm
@@ -44,7 +44,7 @@ int peter_lunh_sum(long number)
 
 
 // function to give first two digits of the given number
-long first_two_digits(long number)
+int first_two_digits(long number)
 {
     while (no_of_digits(number) > 2)
     {
